Set up colorway for automatic color handling
This update implements @helpscout/colorway, which automatically pulls the
latest colour configs for HSDS into the project on start/build.
