+++ src/styles/configs/_color.scss
@@ -1,21 +1,8 @@
 @import 'pack/seed-color-scheme/_index';
+@import './colorway';
 
-// Text
 @include _color(
   (
-    blue:
-      (
-        default: #1292ee,
-        100: #f3fbff,
-        200: #d6edff,
-        300: #a0d4ff,
-        400: #57b0fb,
-        500: #1292ee,
-        600: #0077cc,
-        700: #005ca4,
-        800: #034077,
-        900: #002651
-      ),
     note: (background: #fff7dd),
     text:
       (
