+++ src/styles/configs/_colorway.scss
@@ -1,0 +1,20 @@
+@import 'pack/seed-color-scheme/_index';
+
+// Text
+@include _color(
+  (
+    blue:
+      (
+        default: #1292ee,
+        100: #f3fbff,
+        200: #d6edff,
+        300: #a0d4ff,
+        400: #57b0fb,
+        500: #1292ee,
+        600: #0077cc,
+        700: #005ca4,
+        800: #034077,
+        900: #002651
+      )
+  )
+);
